# docs/swagger.json
{"openapi":"3.0.1","info":{"title":"API Documentation","description":"API Documentation for My Application","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/users/{id}":{"get":{"tags":["User"],"operationId":"getUsersWithRole","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"put":{"tags":["User"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"delete":{"tags":["User"],"operationId":"deleteUser_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/institutions/{id}":{"get":{"tags":["Institution"],"operationId":"getInstitution","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InstitutionResponse"}}}}}},"put":{"tags":["Institution"],"operationId":"updateInstitution","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstitutionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InstitutionResponse"}}}}}},"delete":{"tags":["Institution"],"operationId":"deleteInstitution","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InstitutionResponse"}}}}}}},"/api/donations/{id}":{"get":{"tags":["Donation"],"operationId":"getDonation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DonationResponse"}}}}}},"put":{"tags":["Donation"],"operationId":"updateDonation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DonationUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DonationResponse"}}}}}},"delete":{"tags":["Donation"],"operationId":"deleteDonation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DonationResponse"}}}}}}},"/api/categories/{id}":{"get":{"tags":["Category"],"operationId":"getCategory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryResponse"}}}}}},"put":{"tags":["Category"],"operationId":"updateCategory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryResponse"}}}}}},"delete":{"tags":["Category"],"operationId":"deleteCategory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryResponse"}}}}}}},"/refresh_token":{"post":{"tags":["Authentication"],"operationId":"refreshToken","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/login":{"post":{"tags":["Authentication"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/users/registration":{"post":{"tags":["User"],"operationId":"registration","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegistrationResponse"}}}}}}},"/api/users/recovery/{email}":{"post":{"tags":["User"],"summary":"Sends a token to the specified email address with the option to change the password.","operationId":"postRecoveryPasswordEmail","parameters":[{"name":"email","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmailCheckEmailResponse"}}}}}}},"/api/users/recovery/password":{"get":{"tags":["User"],"summary":"Authenticate the recovery password token to ensure it's validity.","operationId":"getRecoveryPassword","parameters":[{"name":"resetPasswordCheckTokenRequest","in":"query","required":true,"schema":{"$ref":"#/components/schemas/ResetPasswordCheckTokenRequest"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResetPasswordCheckTokenResponse"}}}}}},"post":{"tags":["User"],"summary":"Authenticate the recovery password token to ensure it's validity and then reset the user's password accordingly.","operationId":"postRecoveryPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResetPasswordResponse"}}}}}}},"/api/institutions/add":{"post":{"tags":["Institution"],"operationId":"addInstitution","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstitutionAddRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InstitutionAddResponse"}}}}}}},"/api/donations/add":{"post":{"tags":["Donation"],"operationId":"addDonationUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DonationAddRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DonationAddResponse"}}}}}}},"/api/categories/add":{"post":{"tags":["Category"],"operationId":"addCategory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryAddRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryAddResponse"}}}}}}},"/api/users":{"get":{"tags":["User"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserListResponse"}}}}}}},"/api/users/verification":{"get":{"tags":["User"],"summary":"Verify the user's token for authenticity, then activate the user account.","operationId":"activateUser","parameters":[{"name":"activateUserRequest","in":"query","required":true,"schema":{"$ref":"#/components/schemas/ActivateUserRequest"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ActivateUserResponse"}}}}}}},"/api/users/role/{role}":{"get":{"tags":["User"],"operationId":"getUsersWithRole_1","parameters":[{"name":"role","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserListResponse"}}}}}}},"/api/institutions":{"get":{"tags":["Institution"],"operationId":"getAllInstitutions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InstitutionListResponse"}}}}}}},"/api/donations":{"get":{"tags":["Donation"],"operationId":"getAllDonations","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DonationListResponse"}}}}}}},"/api/categories":{"get":{"tags":["Category"],"operationId":"getAllCategories","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryListResponse"}}}}}}}},"components":{"schemas":{"UserRequest":{"type":"object","properties":{"email":{"maxLength":50,"minLength":0,"type":"string"},"name":{"maxLength":30,"minLength":0,"type":"string"},"lastName":{"maxLength":30,"minLength":0,"type":"string"},"password":{"type":"string"},"enabled":{"type":"boolean"},"roleIdList":{"uniqueItems":true,"type":"array","items":{"type":"integer","format":"int64"}}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Role":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"authority":{"type":"string"}}},"User":{"required":["email","lastName","name","password"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"maxLength":50,"minLength":0,"type":"string"},"name":{"maxLength":30,"minLength":0,"type":"string"},"lastName":{"maxLength":30,"minLength":0,"type":"string"},"password":{"type":"string"},"enabled":{"type":"boolean"},"role":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Role"}},"token":{"type":"string"},"createdAccount":{"type":"string","format":"date-time"},"username":{"type":"string"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"accountNonLocked":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"},"accountNonExpired":{"type":"boolean"}}},"UserResponse":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/User"}}},"InstitutionRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}},"Institution":{"required":["description","name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"}}},"InstitutionResponse":{"type":"object","properties":{"institution":{"$ref":"#/components/schemas/Institution"}}},"DonationAddress":{"type":"object","properties":{"street":{"maxLength":100,"minLength":0,"type":"string"},"city":{"maxLength":100,"minLength":0,"type":"string"},"zipCode":{"maxLength":50,"minLength":0,"type":"string"},"phoneNumber":{"minimum":1,"type":"integer","format":"int64"},"pickUpDate":{"type":"string","format":"date"},"pickUpTime":{"$ref":"#/components/schemas/LocalTime"},"pickUpComment":{"maxLength":500,"minLength":0,"type":"string"}}},"DonationUpdateRequest":{"type":"object","properties":{"quantity":{"minimum":1,"type":"integer","format":"int32"},"categoryIdList":{"type":"array","items":{"type":"integer","format":"int64"}},"institutionId":{"type":"integer","format":"int64"},"donationAddress":{"$ref":"#/components/schemas/DonationAddress"},"receive":{"type":"boolean"},"userId":{"type":"integer","format":"int64"}}},"LocalTime":{"type":"object","properties":{"hour":{"type":"integer","format":"int32"},"minute":{"type":"integer","format":"int32"},"second":{"type":"integer","format":"int32"},"nano":{"type":"integer","format":"int32"}}},"Category":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"Donation":{"required":["category","city","pickUpDate","pickUpTime","quantity","street","zipCode"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"},"category":{"type":"array","items":{"$ref":"#/components/schemas/Category"}},"institution":{"$ref":"#/components/schemas/Institution"},"street":{"maxLength":100,"minLength":0,"type":"string"},"city":{"maxLength":100,"minLength":0,"type":"string"},"zipCode":{"maxLength":50,"minLength":0,"type":"string"},"phoneNumber":{"minimum":1,"type":"integer","format":"int64"},"pickUpDate":{"type":"string","format":"date"},"pickUpTime":{"$ref":"#/components/schemas/LocalTime"},"pickUpComment":{"maxLength":500,"minLength":0,"type":"string"},"receive":{"type":"boolean"},"createdDate":{"type":"string","format":"date"},"createdTime":{"$ref":"#/components/schemas/LocalTime"},"user":{"$ref":"#/components/schemas/User"}}},"DonationResponse":{"type":"object","properties":{"donation":{"$ref":"#/components/schemas/Donation"}}},"CategoryRequest":{"type":"object","properties":{"name":{"type":"string"}}},"CategoryResponse":{"type":"object","properties":{"category":{"$ref":"#/components/schemas/Category"}}},"UserLoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"AuthenticationResponse":{"type":"object","properties":{"access_token":{"type":"string"},"refresh_token":{"type":"string"},"message":{"type":"string"}}},"RegistrationRequest":{"required":["email","firstName","lastName","password","repeatPassword"],"type":"object","properties":{"email":{"maxLength":50,"minLength":0,"type":"string"},"firstName":{"maxLength":30,"minLength":0,"type":"string"},"lastName":{"maxLength":30,"minLength":0,"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"pattern":"^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[!*@#$%^&+=])(?=\\S+$).{8,}$","type":"string"},"repeatPassword":{"type":"string"}}},"RegistrationResponse":{"type":"object","properties":{"successful":{"type":"boolean"},"message":{"type":"string"},"registrationRequest":{"$ref":"#/components/schemas/RegistrationRequest"}}},"EmailCheckEmailResponse":{"type":"object","properties":{"successful":{"type":"boolean"},"message":{"type":"string"}}},"ResetPasswordRequest":{"type":"object","properties":{"token":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"pattern":"^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[!*@#$%^&+=])(?=\\S+$).{8,}$","type":"string"},"repeatPassword":{"type":"string"}}},"ResetPasswordResponse":{"type":"object","properties":{"successful":{"type":"boolean"},"message":{"type":"string"}}},"InstitutionAddRequest":{"required":["description","name"],"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}},"InstitutionAddResponse":{"type":"object","properties":{"institutionAddRequest":{"$ref":"#/components/schemas/InstitutionAddRequest"}}},"DonationAddRequest":{"required":["categoryIdList","city","institutionId","pickUpDate","pickUpTime","quantity","street","userId","zipCode"],"type":"object","properties":{"quantity":{"type":"integer","format":"int32"},"categoryIdList":{"type":"array","items":{"type":"integer","format":"int64"}},"institutionId":{"type":"integer","format":"int64"},"street":{"maxLength":100,"minLength":0,"type":"string"},"city":{"maxLength":100,"minLength":0,"type":"string"},"zipCode":{"maxLength":50,"minLength":0,"type":"string"},"phoneNumber":{"minimum":1,"type":"integer","format":"int64"},"pickUpDate":{"type":"string","format":"date"},"pickUpTime":{"$ref":"#/components/schemas/LocalTime"},"receive":{"type":"boolean"},"createdDate":{"type":"string","format":"date"},"createdTime":{"$ref":"#/components/schemas/LocalTime"},"pickUpComment":{"maxLength":500,"minLength":0,"type":"string"},"userId":{"type":"integer","format":"int64"}}},"DonationAddResponse":{"type":"object","properties":{"donationAddRequest":{"$ref":"#/components/schemas/DonationAddRequest"}}},"CategoryAddRequest":{"required":["name"],"type":"object","properties":{"name":{"type":"string"}}},"CategoryAddResponse":{"type":"object","properties":{"categoryAddRequest":{"$ref":"#/components/schemas/CategoryAddRequest"}}},"UserListResponse":{"type":"object","properties":{"userList":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}},"ActivateUserRequest":{"type":"object","properties":{"token":{"type":"string"}}},"ActivateUserResponse":{"type":"object","properties":{"successful":{"type":"boolean"},"message":{"type":"string"}}},"ResetPasswordCheckTokenRequest":{"type":"object","properties":{"token":{"type":"string"}}},"ResetPasswordCheckTokenResponse":{"type":"object","properties":{"successful":{"type":"boolean"},"message":{"type":"string"}}},"InstitutionListResponse":{"type":"object","properties":{"institutionList":{"type":"array","items":{"$ref":"#/components/schemas/Institution"}}}},"DonationListResponse":{"type":"object","properties":{"donationList":{"type":"array","items":{"$ref":"#/components/schemas/Donation"}}}},"CategoryListResponse":{"type":"object","properties":{"categoryList":{"type":"array","items":{"$ref":"#/components/schemas/Category"}}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT auth description","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}